List all categories works as before
