--- src/Controllers/Vulnerabilities/Categories/GetVulnerabilityParentCategoriesController.php
@@ -1,19 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Reconmap\Controllers\Vulnerabilities\Categories;
-
-use Psr\Http\Message\ServerRequestInterface;
-use Reconmap\Controllers\Controller;
-use Reconmap\Repositories\VulnerabilityCategoryRepository;
-
-class GetVulnerabilityParentCategoriesController extends Controller
-{
-    public function __construct(private VulnerabilityCategoryRepository $repository)
-    {
-    }
-
-    public function __invoke(ServerRequestInterface $request): array
-    {
-        return $this->repository->findAllWithoutChildren();
-    }
-}
