+++ src/Repositories/VulnerabilityCategoryRepository.php
@@ -40,9 +40,13 @@ class VulnerabilityCategoryRepository extends MysqlRepository implements Deletab
         return $this->deleteByTableId(self::TABLE_NAME, $id);
     }
     
-    public function search(VulnerabilityCategoryCriteria $searchCriteria, ?PaginationRequestHandler $paginator = null, ?string $orderBy = 'name ASC'): array
+    public function search(VulnerabilityCategoryCriteria $searchCriteria, bool $fullSearch = false, ?PaginationRequestHandler $paginator = null, ?string $orderBy = 'id ASC'): array
     {
         $queryBuilder = $this->getBaseSelectQueryBuilder();
+        if ($fullSearch)
+        {
+            $queryBuilder = $this->getSelectWithParentQueryBuilder();
+        }
         return $this->searchAll($queryBuilder, $searchCriteria, $paginator, $orderBy);
     }
 
@@ -50,6 +54,14 @@ class VulnerabilityCategoryRepository extends MysqlRepository implements Deletab
     {
         $queryBuilder = new SelectQueryBuilder('vulnerability_category');
         $queryBuilder->setColumns('id, name, description, parent_id');
+        return $queryBuilder;
+    }
+
+    protected function getSelectWithParentQueryBuilder(): SelectQueryBuilder
+    {
+        $queryBuilder = new SelectQueryBuilder('vulnerability_category vc');
+        $queryBuilder->setColumns('vc.id, vc.name, vc.description, vc.parent_id, parent_vc.name as parent_name');
+        $queryBuilder->addJoin('LEFT JOIN vulnerability_category parent_vc ON (parent_vc.id = vc.parent_id)');
         return $queryBuilder;
     }
 
@@ -63,6 +75,6 @@ class VulnerabilityCategoryRepository extends MysqlRepository implements Deletab
     public function findAll(): array
     {
         $searchCriteria = new VulnerabilityCategoryCriteria();
-        return $this->search($searchCriteria);
+        return $this->search($searchCriteria, true);
     }
 }
