--- tests/Controllers/Vulnerabilities/Categories/GetVulnerabilityParentCategoriesControllerTest.php
@@ -1,27 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Reconmap\Controllers\Vulnerabilities\Categories;
-
-use PHPUnit\Framework\TestCase;
-use Psr\Http\Message\ServerRequestInterface;
-use Reconmap\Repositories\VulnerabilityCategoryRepository;
-
-class GetVulnerabilityParentCategoriesControllerTest extends TestCase
-{
-    public function testHappyPath()
-    {
-        $mockCategory = ['title' => 'foo'];
-
-        $mockRequest = $this->createMock(ServerRequestInterface::class);
-
-        $mockRepository = $this->createMock(VulnerabilityCategoryRepository::class);
-        $mockRepository->expects($this->once())
-            ->method('findAllWithoutChildren')
-            ->willReturn([$mockCategory]);
-
-        $controller = new GetVulnerabilityParentCategoriesController($mockRepository);
-        $response = $controller($mockRequest);
-
-        $this->assertEquals([$mockCategory], $response);
-    }
-}
